mapa

+++ cub3D_bonus/pudge_hook/ft_hook.c
@@ -7,8 +7,8 @@
 int		key_hook(t_all *xlm)
 {
 	xlm->yyy += SPEED * xlm->player.dir_y;
-	if (!xlm->param.save)
-		mlx_mouse_move(xlm->win, (int)(xlm->param.width / 2), (int)(xlm->param.height / 2));
+//	if (!xlm->param.save)
+//		mlx_mouse_move(xlm->win, (int)(xlm->param.width / 2), (int)(xlm->param.height / 2));
 	ft_move(xlm);
 	xlm->img.img = mlx_new_image(xlm->mlx, xlm->param.width, xlm->param.height);
 	xlm->img.adr = mlx_get_data_addr(xlm->img.img, &xlm->img.bpp, &xlm->img.line_len, &xlm->img.iend);
@@ -23,7 +23,9 @@ int		key_hook(t_all *xlm)
 	{
 		xlm->img.img ? mlx_destroy_image(xlm->mlx, xlm->img.img) : 0;
 		mlx_destroy_window(xlm->mlx, xlm->win);
-		system("killall afplay");
+		if (xlm->param.lvl == 1)
+			system("killall afplay");
+		write(1, "Your died.\n", 11);
 		ft_error(xlm, xlm->param.valid);
 		exit(0);
 	}
