+++ cub3D_bonus/pudge_hook/ft_move.c
@@ -109,7 +109,8 @@ void	ft_move(t_all *xlm)
 	{
 		xlm->img.img ? mlx_destroy_image(xlm->mlx, xlm->img.img) : 0;
 		mlx_destroy_window(xlm->mlx, xlm->win);
-		system("killall afplay");
+		if (xlm->param.lvl == 1)
+			system("killall afplay");
 		ft_error(xlm, xlm->param.valid);
 		exit(0);
 	}
